// src/react/collections/Form/lib/withFormikField.types.ts
export type FormikFieldComponentRenderProps<P, ValueType, DisplayedValue = ValueType> = {
  /** Meta props generated by Formik context */
  meta: FieldMetaProps<ValueType>;

  /** All default component props passed while rendering */
  rest: {
    /** On Blur Handler */
    onBlur: (...args: any[]) => void;
    /** On Change Handler */
    onChange: (...args: any[]) => void;
    /** The field name */
    name: string;
    /** The field value */
    value: DisplayedValue
  } & P,

  /** The field state at render time */
  state: FieldState;
};


/* --------
 * Computed Field State generated by the withFormikField
 * HOC Component Function
 * -------- */
type FieldState = SharedComponentStateProps & {
  /** Get if form is currently submitting */
  isSubmitting: boolean;

  /** Validation Message */
  message: string | undefined;
};